mentor_certs: Default certificate fetching
Created a ModeManager method for Certificate to fetch the default one
for display. Did this by playing with the defaulf model record ordering
so it took some migration file changes to properly make it happen.

Along the way also made sure pytest runs in debug mode so test data is
available.